Move SingleSTepNByteToMessageDecoder stuff to separate file
